+++ mod-batteries.ts
@@ -1,0 +1,3 @@
+export { HtmlShell } from "./src/HtmlShell.ts";
+export { byoHat, Hat } from "./src/miscellaneous/Hat.ts";
+export * from "./src/minifyCss.ts";