+++ src/getEnv.ts
@@ -15,4 +15,8 @@ export const IS_DEBUG = DEBUG ? DEBUG !== "false" : false;
 // a bit shameless for Deno use cases, but UX is above all
 export const NODE_ENV = getEnv("NODE_ENV");
 export const IS_PROD = NODE_ENV === "production";
-export const IS_DEV = NODE_ENV !== "production";
+
+const PREACT_DEVTOOLS = getEnv("PREACT_DEVTOOLS");
+export const PREACT_DEVTOOLS_ENABLED = PREACT_DEVTOOLS
+  ? PREACT_DEVTOOLS !== "false"
+  : false;