+++ src/ssr.ts
@@ -1,5 +1,5 @@
 import type { FunctionComponent, VNode } from "preact";
-import type { CP } from "./types.ts";
+import type { CP, SsrAdditionalOptions } from "./types.ts";
 import { renderToStringAsync } from "preact-render-to-string";
 import { withHtmlShell } from "./HtmlShell.ts";
 import { cleanCss } from "./minifyCss.ts";
@@ -30,6 +30,7 @@ export const byoHtml = async (component: VNode): Promise<string> => {
 
 const ssrVNode = async (
   node: VNode,
+  opts: SsrAdditionalOptions = {},
 ): Promise<string> => {
   const componentName = typeof node.type === "string"
     ? node.type
@@ -40,12 +41,14 @@ const ssrVNode = async (
     componentProps: node.props,
     component: node,
     css: cleanCss(componentCss),
+    opts,
   }));
 };
 
 const ssrTsxComponent = async <P extends CP>(
   component: FunctionComponent<P>,
   props: P | null = null,
+  opts: SsrAdditionalOptions = {},
 ): Promise<string> => {
   const componentName = component.name;
 
@@ -61,6 +64,7 @@ const ssrTsxComponent = async <P extends CP>(
     // component: h<CP>(componentName, props),
     component: componentVNode,
     css: cleanCss(componentCss),
+    opts,
   }));
 };
 
@@ -69,15 +73,17 @@ const ssrTsxComponent = async <P extends CP>(
  * then renders everything out as a string
  * @param renderTarget - a Preact VNode or a TSX component
  * @param componentProps - the props to pass to the component (only applicable when passing a TSX component as `renderTarget`)
+ * @param opts additional options (e.g. the meta tags useful for SEO purposes)
  * @returns - the rendered HTML as a string
  * @TODO update the README example
  */
 export const ssr = <P extends CP>(
   renderTarget: FunctionComponent<P> | VNode,
   componentProps: P | null = null,
+  opts: SsrAdditionalOptions = {},
 ): Promise<string> => {
   // @TODO see if there's a better way to check if it's a VNode
   const isVNode = typeof renderTarget === "object" && "type" in renderTarget;
-  if (isVNode) return ssrVNode(renderTarget);
-  return ssrTsxComponent(renderTarget, componentProps);
+  if (isVNode) return ssrVNode(renderTarget, opts);
+  return ssrTsxComponent(renderTarget, componentProps, opts);
 };