+++ src/types.ts
@@ -1,4 +1,13 @@
 import type { ComponentProps, ComponentType } from "preact";
+import type { VNode } from "preact";
+
+export type SsrAdditionalOptions = {
+  /**
+   * the content of all tags found in the html <head> element
+   * such as `<meta>`, `<title>`, and so on
+   */
+  headContent?: VNode;
+};
 
 /**
  * @ignore